Mejorada la estructura de clases
